mds: only update change_attr and btime when client sets appropriate feature flags
The kernel client lags the userland code a bit, and feature support for
addr2 is not quite ready. Still, we want to allow the client to set the
new flags field in a cap request before then so it can get better fsync
performance.

When we go to update the cap fields, grab the features from the peer,
and verify that the appropriate flags are set before we apply updates
to the btime and change_attr.

Signed-off-by: Jeff Layton <jlayton@redhat.com>